+++ dantro/data_loaders/load_text.py
@@ -1,0 +1,30 @@
+"""Defines a loader mixin to load plain text objects"""
+
+from ..containers import StringContainer
+from ._tools import add_loader
+
+# -----------------------------------------------------------------------------
+
+class TextLoaderMixin:
+    """Supplies functionality to load plain text"""
+
+    @add_loader(TargetCls=StringContainer)
+    def _load_plain_text(filepath: str, *, TargetCls: type,
+                         **load_kwargs) -> StringContainer:
+        """Loads the content of a plain text file back into a StringContainer.
+        
+        Args:
+            filepath (str): Where the *.txt file is located
+            TargetCls (type): The class constructor
+            **load_kwargs: Passed on to open, see there for kwargs
+        
+        Returns:
+            StringContainer: The reconstructed StringContainer
+        """
+        with open(filepath, **load_kwargs) as f:
+            data = f.read()
+
+        return TargetCls(data=data, attrs=dict(filepath=filepath))
+
+    # Also make it available under plain 'text'
+    _load_text = _load_plain_text