+++ tests/test_data_mngr.py
@@ -11,13 +11,15 @@ import pytest
 
 import dantro.base
 from dantro.containers import (NumpyDataContainer, XrDataContainer,
-                               ObjectContainer, PassthroughContainer)
+                               ObjectContainer, PassthroughContainer,
+                               StringContainer)
 from dantro.groups import OrderedDataGroup
 from dantro.mixins import Hdf5ProxySupportMixin
 import dantro.data_mngr
 from dantro.data_loaders import (YamlLoaderMixin, PickleLoaderMixin,
                                  Hdf5LoaderMixin, NumpyLoaderMixin,
                                  XarrayLoaderMixin,
+                                 TextLoaderMixin,
                                  AllAvailableLoadersMixin)
 from dantro.tools import write_yml
 
@@ -34,6 +36,10 @@ class DataManager(YamlLoaderMixin, dantro.data_mngr.DataManager):
     # A (bad) load function for testing
     def _load_bad_loadfunc(self):
         pass
+
+class TextDataManager(TextLoaderMixin, DataManager):
+    """A data manager that is able to load text files"""
+    pass
 
 class PklDataManager(PickleLoaderMixin, DataManager):
     """A data manager that is able to load pickled files"""
@@ -105,6 +111,22 @@ def data_dir(tmpdir) -> str:
 def dm(data_dir) -> DataManager:
     """Returns a DataManager without load configuration"""
     return DataManager(data_dir, out_dir=None)
+
+@pytest.fixture
+def text_dm(data_dir) -> TextDataManager:
+    """Manager with test data for TextLoaderMixin"""
+    # Create a sundirectory for the text data
+    text_dir = data_dir.mkdir("text_data")
+
+    # Define a string to dump
+    to_dump = "This is a test string \n with two lines\n"
+
+    # save the file
+    with open(text_dir.join("test_string.txt"), mode="w") as f:
+        f.write(to_dump)
+
+    return TextDataManager(data_dir, out_dir=None)
+    
 
 @pytest.fixture
 def pkl_dm(data_dir) -> PklDataManager:
@@ -693,6 +715,20 @@ def test_target_path(dm):
     assert 'barfoo_group/lamo' in dm
     assert 'barfoo_group/also_lamo' in dm
 
+
+# TextLoaderMixin tests -------------------------------------------------------
+
+def test_text_loader(text_dm):
+    """Test the plain text loader"""
+    text_dm.load('text_data', loader='text', glob_str="text_data/*.txt")
+
+    # Check that the plain text data is loaded and of expected type
+    text_data = text_dm['text_data']
+
+    for name, cont in text_data.items():
+        assert isinstance(cont, StringContainer)
+    
+        assert cont.data == "This is a test string \n with two lines\n"
 
 # PickleLoaderMixin tests -----------------------------------------------------